+++ src/components/views/ViewMap.vue
@@ -176,6 +176,7 @@ import type { UserBookmarkPreset } from '~/utils/backend/handlers/bookmarks';
 import { showBookmark } from '~/composables/fetchers';
 import { fromLonLat, toLonLat, transformExtent } from 'ol/proj';
 import { useRadarError } from '~/composables/errors';
+import {getPilotTrueAltitude} from "~/utils/shared/vatsim";
 
 defineProps({
     mode: {
@@ -456,7 +457,7 @@ useUpdateInterval(() => {
     if (!user) return;
 
     setUserLocalSettings({
-        vatglassesLevel: Math.round(user.altitude / 500) * 5,
+        vatglassesLevel: Math.round(getPilotTrueAltitude(user) / 500) * 5,
     });
 });